+++ js/index.js
@@ -1,0 +1,5 @@
+import BookForm from "../modules/book.js";
+import {setTimeDate} from "../modules/ui.js";
+
+new BookForm();
+setTimeDate();
